pyproject.toml
[project]
name = "cozy-voice"
version = "0.1.0"
description = "CozyEngineV2 语音通道：STT/TTS/Realtime Provider 抽象 + FastRTC 原型"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "openai>=1.50",
    "tencentcloud-sdk-python-tts>=3.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: 需启动 CozyVoice 服务的集成测试",
    "integration_real_voice: 需真 OpenAI/腾讯 API key 的真调测试",
]
